sync-github-releases: rename github-utils.ts to github.ts
The file is the GitHub REST API client — a clear abstraction. The "-utils"
suffix described nothing (a utils file is a junk drawer); name it for what it is.
